backend/knowledge_agent/requirements.txt
a2a-sdk==0.2.16
click>=8.1.8
httpx>=0.28.1
langchain-google-genai==2.1.3
google-ai-generativelanguage==0.6.17
langgraph==0.6.3
langchain-openai ==0.3.28
pydantic>=2.10.6
python-dotenv>=1.1.0
uvicorn>=0.34.2
langchain-mcp-adapters
rapidfuzz
elasticsearch==7.12.1
mysql-connector-python